Compatibility with 4.10 panel
The xkb plugin now always occupies one row only (xfce_panel_plugin_set_small).
If the panel has only one row, the plugin requests a rectangular shape
(just as it used to); with more than one panel row, the plugin
will be the "standard" square so as to better blend in with other
plugins at the expense of having smaller display area.

Bug 8762