Ignore xkl config change callbacks caused by our own config changes.
libxklavier emits X-config-changed signal from within
xkl_config_rec_activate before it gets a chance to fully update
its internal state so that xkl_config_rec_get_from_server returns
the *old* xkl config when invoked from the callback. This, in turn,
updates xkb-plugin's internal state to the old xkl config and causes it
go out of sync with the real xkl state.

Since X-config-changed signal is synchronous, the easiest work around
is to simply ignore the xkl config change notification when we know
we're in the middle of updating xkl config.