Add optional notification support (Fixes #48)
Add support for showing notifications on layout changes. At build time,
this depends on the availability of libnotify. At run time, it is
controlled by an xfconf property / setting in the properties dialog.

The notifications can get quite spammy when the layout is managed per
application or per window, which is why it feels important that they can
be disabled.